+++ packages/web/src/components/layouts/Layout.tsx
@@ -8,7 +8,7 @@ export default function Layout() {
       <MarqueeHeader />
       <div className="flex flex-1 overflow-hidden">
         <Sidebar />
-        <main className="flex-1 overflow-y-auto p-6">
+        <main className="flex-1 overflow-y-auto p-5 lg:p-6">
           <Outlet />
         </main>
       </div>